First pass at functionality for RecentSearches
Program is designed to scrape Twitter's API via
keyword for tweets in the last 7 days matching the keyword
or sentence passed in by the user. The program
saves a local txt file of raw data and
then process the raw data to only include the
userID and their corresponding tweet.

This allows the programmer to easily access data
and do what they choose with it. Plans for next steps
are TBD.

Currently this does not have a UI, however
one is intended for the future.

Author: Michael J. Isabella
Date: 11-18-22