Cleaned up and realized potential for refactor
Currently recursive filter targets JSON by index value
rather than by keyword. This has a potential to break
if Twitter changes how their API returns raw data.

The fix would be simple if they do it with the current
amount of data Twitter sends, but if that changes
it could become more complicated. Something to keep
in mind as you look for opportunities to refactor.