Refactored scripts to reduce time and space complexities
Previously the 'combineTweets-node.js' script utilized a double
nested loops which increased the time complexity to 0(n^2).
Instead the functionality was refactored to reference object keys
directly to reduce the time complexity to O(n).

Furthermore, after further review, the 'combineTweets-node.js'
script was not summarizing all of the past searches into one
document, but instead was only writing the first search and
ignoring the remaining results. This was resolved by swapping out
writeFile with appendFile.

Previously the 'searchTwitter-node.js' script broke the DRY
principle, so refactorization of code cut down on code
for readability and simplified the program flow to adhere
to the DRY principle.

Overall the program is more readable, faster and more scalable
so that unnecessary variables aren't created.

Michael J. Isabella - Dec. 27th, 2022